ProductionFiles/requirements.txt
botocore==1.42.13
fastapi==0.125.0
from_root==1.3.0
mypy_boto3_s3==1.42.10
numpy==1.26.4
pandas==2.3.3
scikit_learn==1.7.2
uvicorn==0.38.0
